backend/measurement_system/requirements.txt
flask
opencv-python-headless
numpy
scipy
imutils
gunicorn
pillow
